Move `DuplicateCache` to reduce diff
