+++ beacon_node/network/src/network_beacon_processor/mod.rs
@@ -7,7 +7,7 @@ use beacon_chain::{
 };
 use beacon_chain::{BeaconChainTypes, NotifyExecutionLayer};
 use beacon_processor::{
-    work_reprocessing_queue::ReprocessQueueMessage, GossipAggregatePackage,
+    work_reprocessing_queue::ReprocessQueueMessage, DuplicateCache, GossipAggregatePackage,
     GossipAttestationPackage, Work, WorkEvent as BeaconWorkEvent, MAX_SCHEDULED_WORK_QUEUE_LEN,
     MAX_WORK_EVENT_QUEUE_LEN,
 };
@@ -16,10 +16,8 @@ use lighthouse_network::{
     rpc::{BlocksByRangeRequest, BlocksByRootRequest, LightClientBootstrapRequest, StatusMessage},
     Client, MessageId, NetworkGlobals, PeerId, PeerRequestId,
 };
-use parking_lot::Mutex;
 use slog::{debug, Logger};
 use slot_clock::ManualSlotClock;
-use std::collections::HashSet;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
@@ -45,55 +43,6 @@ pub(crate) const FUTURE_SLOT_TOLERANCE: u64 = 1;
 pub enum InvalidBlockStorage {
     Enabled(PathBuf),
     Disabled,
-}
-
-/// A handle that sends a message on the provided channel to a receiver when it gets dropped.
-///
-/// The receiver task is responsible for removing the provided `entry` from the `DuplicateCache`
-/// and perform any other necessary cleanup.
-pub struct DuplicateCacheHandle {
-    entry: Hash256,
-    cache: DuplicateCache,
-}
-
-impl Drop for DuplicateCacheHandle {
-    fn drop(&mut self) {
-        self.cache.remove(&self.entry);
-    }
-}
-
-/// A simple  cache for detecting duplicate block roots across multiple threads.
-#[derive(Clone, Default)]
-pub struct DuplicateCache {
-    inner: Arc<Mutex<HashSet<Hash256>>>,
-}
-
-impl DuplicateCache {
-    /// Checks if the given block_root exists and inserts it into the cache if
-    /// it doesn't exist.
-    ///
-    /// Returns a `Some(DuplicateCacheHandle)` if the block_root was successfully
-    /// inserted and `None` if the block root already existed in the cache.
-    ///
-    /// The handle removes the entry from the cache when it is dropped. This ensures that any unclean
-    /// shutdowns in the worker tasks does not leave inconsistent state in the cache.
-    pub fn check_and_insert(&self, block_root: Hash256) -> Option<DuplicateCacheHandle> {
-        let mut inner = self.inner.lock();
-        if inner.insert(block_root) {
-            Some(DuplicateCacheHandle {
-                entry: block_root,
-                cache: self.clone(),
-            })
-        } else {
-            None
-        }
-    }
-
-    /// Remove the given block_root from the cache.
-    pub fn remove(&self, block_root: &Hash256) {
-        let mut inner = self.inner.lock();
-        inner.remove(block_root);
-    }
 }
 
 /// Provides an interface to a `BeaconProcessor` running in some other thread.
